Cleaning up.
git-svn-id: svn+ssh://cv077/home/ueshiba/svnroot/trunk/src@1602 3ef1b8b6-fb57-41bf-88d9-9a5bd220bf38

+++ Testprog/SURF/MatchImage.cc
@@ -9,7 +9,7 @@ namespace TU
 *  class MatchImage							*
 ************************************************************************/
 MatchImage&
-MatchImage::copy(const Image<u_char>& image, int u0, int v0)
+MatchImage::copy(const Image<u_char>& image, size_t u0, size_t v0)
 {
     for (size_t v = 0; v < image.height(); ++v)
 	for (size_t u = 0; u < image.width(); ++u)
